feat: move scrolling to viewport
